Send token with requests

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,15 +1,23 @@
 import useWorkoutContext from "../hooks/useWorkoutContext";
+import useAuthContext from '../hooks/useAuthContext'
 import { formatDistanceToNow } from "date-fns";
 import TrashIcon from "../assets/trash-bin.png";
 import PropTypes from "prop-types";
 
 const WorkoutDetails = ({workout}) => {
   const { dispatch } = useWorkoutContext();
+  const { user } = useAuthContext()
 
   const handleDelete = async () => {
+    if (!user) {
+      return
+    }
     // we didn't put localhost:4000 in the fetch url because we set the proxy in package.json
     const response = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      },
     });
     const data = await response.json();
     if (response.ok) {
